feat: introduce selector field to requests (#89)
## This PR

Introduces an optional filtering field to grpc requests. 

The intended usage of the new field: `selector `

```
Optional: A selector for the flag configuration request. The server implementation may utilize this to select
flag configurations from a collection, select the source of the flag or combine this to any desired underlying
filtering mechanism.
   ex:- selector: 'source=database,app=weatherapp'
```
For example, flagd could utilize this filtering field to point out the
K8s CRD when connecting to kube-proxy [1]

[1] - https://github.com/open-feature/flagd/pull/495

---------

Signed-off-by: Kavindu Dodanduwa <kavindudodanduwa@gmail.com>