feat: add metadata endpoint (#113)
This PR adds a new method and message definitions for retrieving
metadata from a sync provider. What exactly the attributes of the
metadata response are is up to the implementation of the sync service,
as the response for the `GetMetadata` requests is a string map
containing arbitrary keys. Examples for those attributes could be the
name of the environment or the kubernetes cluster the service is running
in

Signed-off-by: Florian Bacher <florian.bacher@dynatrace.com>