feat!: struct for metadata for consistency and power (#130)
This PR alters the _new_ schemas to use maps consistently for all
metadata.

Previously, the `getMetadata` RPC was returning a custom key/val
response, which is the old proto2 way of doing maps.

Additionally, for consistency, I have changed the flag metadata to
`map<string, string>` as well. This is a "more" breaking change, but
it's consistent and will allow for much stronger typing. If there's a
need for more flexibility in flag metadata, we can revert.

We can avoid this change if the breaking is a concern, but I think it's
save to assume nobody but us are using this new proto.

---------

Signed-off-by: Todd Baert <todd.baert@dynatrace.com>