fix: revert full schema versioning (#166)
Reverts https://github.com/open-feature/flagd-schemas/pull/163. Though
there's a lot of benefits, the maintenance burden here is too high.

It means we will need to update the schema version everywhere,
frequently. For example, we'd have to update it in places like this:
https://github.com/open-feature/java-sdk-contrib/blob/ae21165c2bd65cad325e66e8373fe38da8a98baf/providers/flagd/src/main/java/dev/openfeature/contrib/providers/flagd/resolver/process/model/FlagParser.java#L46-L47
.

I think this will be too error prone and too much of a maintenance
burden for little benefit. I'm reverting back to only major
version-based versioning.

Signed-off-by: Todd Baert <todd.baert@dynatrace.com>