feat: Allow map configurations for flagd and further improvements (#190)
Adding additional schema for import into flagd (called it `flagdConfig`
- please advice for better name) as we only want to use it for flagd but
not the providers. The ADR speaks of a name `flagsArray` which i do
think holds too much implementation detail, and is not as descriptive.
Now we have a `flagdConfig` and a `providerConfig` in our schema repo.

The new schema for flags is not the root object, so we are not
introducing this format immediately, and also all the providers
currently do utilize the root object for validation.

Some minor improvements, like generating a `AnyFlag` object to easier
reference any kind of flag with all variants was also added.

All objects are now part of the definitions, and will be simple
referenced when needed (could be even further improved)

closes: https://github.com/open-feature/flagd-schemas/issues/189

EDIT BY @toddbaert :

Please see:
https://github.com/open-feature/flagd-schemas/pull/190#issuecomment-3211822593

---------

Signed-off-by: Simon Schrottner <simon.schrottner@dynatrace.com>
Signed-off-by: Todd Baert <todd.baert@dynatrace.com>
Co-authored-by: Todd Baert <todd.baert@dynatrace.com>